requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
passlib[bcrypt]
python-jose[cryptography]   
PyJWT                        
pydantic
python-dotenv
redis>=5
aioboto3                     
pillow                       
httpx                        
python-multipart             
